ProdigyInfoTech-WD-Task-02
This is the Second Task that i have been assigned as a Web Development Intern at Prodigy_InfoTech. The Task requirement was to create a interactive stopwatch using javascript with functionalities of Start/Stop and lap count with Reset. I was supposed to style the website and structure it properly using html and css.